refactor: CommonController 삭제 및 기능 분배
* 기존 CommonController의 기능을 DriverController와 ParentController로 분배
